// types/tracking/user/EnrichedCustomerData.ts
// Path: types/tracking/user/EnrichedCustomerData.ts

export type EnrichedCustomerData = {
  email?: string | undefined
  phone?: string | undefined
  externalId?: string | undefined
  firstName?: string | undefined
  lastName?: string | undefined
  city?: string | undefined
  state?: string | undefined
  zip?: string | undefined
  countryCode?: string | undefined
  clientIp?: string | undefined
  userAgent?: string | undefined
  fbp?: string | undefined
  fbc?: string | undefined
}